requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.6.3
attrs==24.2.0
cachetools==4.2.2
certifi==2024.8.30
frozenlist==1.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
nest-asyncio==1.6.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyQRCode==1.2.1
python-dotenv==1.0.1
python-telegram-bot==20.0
pytz==2024.2
rfc3986==1.5.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.17.1
